show a alert if user agree the delete the user

+++ src/components/SingleCoffee.jsx
@@ -1,16 +1,37 @@
-
+import Swal from 'sweetalert2'
+import withReactContent from 'sweetalert2-react-content'
 import PropTypes from 'prop-types';
 
 const SingleCoffee = ({ coffee }) => {
-    const { _id, name, chef, supplier, taste, category, details, photo } = coffee;
+    const MySwal = withReactContent(Swal)
+    const { _id, name, chef, taste, photo } = coffee;
     const handleDeleteCoffee = id => {
-        fetch(`http://localhost:5000/coffee/${id}`, {
-            method: "DELETE"
+        MySwal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                fetch(`http://localhost:5000/coffee/${id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data);
+                        MySwal.fire(
+                            'Deleted!',
+                            'Your file has been deleted.',
+                            'success'
+                        )
+                    })
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-            })
+
     }
     return (
         <div className='rounded-lg'>
